+++ app/src/main/java/com/github/kpdandroid/ui/MainNavDestinations.kt
@@ -1,0 +1,9 @@
+package com.github.kpdandroid.ui
+
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.AccountBox
+import androidx.compose.ui.graphics.vector.ImageVector
+
+enum class MainNavDestinations(val title: String, val icon: ImageVector) {
+    CAMERA_ANALYSIS("Camera analysis", Icons.Default.AccountBox)
+}
